dm: Remove header vmm.h
vmm.h has some unused definitions.

Keep the useful definitions and delete vmm.h

Tracked-On: #6282
Acked-by: Wang, Yu1 <yu1.wang@intel.com>
