+++ devicemodel/include/mem.h
@@ -28,6 +28,8 @@
 
 #ifndef _MEM_H_
 #define	_MEM_H_
+#include "types.h"
+#include "vhm_ioctl_defs.h"
 
 struct vmctx;
