+++ devicemodel/include/vmmapi.h
@@ -32,8 +32,9 @@
 #include <sys/param.h>
 #include <uuid/uuid.h>
 #include "types.h"
-#include "vmm.h"
 #include "macros.h"
+#include "pm.h"
+#include "vhm_ioctl_defs.h"
 
 /*
  * API version for out-of-tree consumers for making compile time decisions.
